CMake refactor. and example fixing

+++ src/modules/Utils/CopyArray_Double.c
@@ -2,7 +2,7 @@
 #include "Utilities.h"
 
 double* copy_array_double(const double arr[], size_t size) {
-  double* buffer = malloc(size * sizeof(int));
+  double* buffer = malloc(size * sizeof(double));
   Null_Pointer_Guard(buffer, __FILE__, __LINE__);
   for (size_t i = 0; i < size; i++) {
     buffer[i] = arr[i];
